myapp/serializers.py
from myapp.models import User  # 我只撿了model裡面的一個小傢伙User
from rest_framework import serializers


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'
